backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
python-dotenv==1.0.0
aiosqlite==0.19.0
sqlalchemy[asyncio]==2.0.25
groq==0.4.2
dnspython==2.4.2
slowapi==0.1.9
python-multipart==0.0.6
email-validator==2.1.0